+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## 1.1.1 (not released)
 
 - Fixes an issue with Java 11 and SSL handshaking.
+- The terminology for relationships can now be customised.
 
 ## 1.1.0 (8th November 2018)